zima-nextjs/requirements.txt
flask==3.0.0
pyyaml==6.0.1
click==8.1.7
rich==13.7.0
gitpython==3.1.40
python-dotenv==1.0.0
requests==2.31.0
psutil>=5.9.0
